+++ common/src/datasets.rs
@@ -2,14 +2,14 @@ use std::u32;
 
 #[derive(Debug, Clone)]
 pub struct DataSet {
-    pub features: Vec<Vec<f32>>,
-    pub labels: Vec<f32>,
+    pub features: Vec<Vec<f64>>,
+    pub labels: Vec<f64>,
 }
 
 #[derive(Debug, Clone)]
 pub struct MultiTargetDataSet {
-    pub features: Vec<Vec<f32>>,
-    pub labels: Vec<Vec<f32>>,
+    pub features: Vec<Vec<f64>>,
+    pub labels: Vec<Vec<f64>>,
     pub indices: Vec<usize>,
 }
