+++ decision-tree/src/utils/calculations/calculations_gini.rs
@@ -1,11 +1,11 @@
 use crate::class_counter::ClassCounter;
 
 pub(super) fn calculate_loss(
-    number_of_rows: f32,
-    true_rows_count: f32,
+    number_of_rows: f64,
+    true_rows_count: f64,
     class_counts_left: &ClassCounter,
     class_counts_right: &ClassCounter,
-) -> f32 {
+) -> f64 {
     let false_rows_count = number_of_rows - true_rows_count;
     let gini_left = calculate_gini(class_counts_left, false_rows_count);
     let gini_right = calculate_gini(class_counts_right, true_rows_count);
@@ -14,11 +14,11 @@ pub(super) fn calculate_loss(
     result
 }
 
-pub(super) fn calculate_gini(class_counts: &ClassCounter, number_of_rows: f32) -> f32 {
-    let impurity: f32 = 1.0;
-    let mut reduction: f32 = 0.0;
+pub(super) fn calculate_gini(class_counts: &ClassCounter, number_of_rows: f64) -> f64 {
+    let impurity: f64 = 1.0;
+    let mut reduction: f64 = 0.0;
     class_counts.counts.iter().for_each(|class_count| {
-        let probability_i = *class_count as f32 / number_of_rows;
+        let probability_i = *class_count as f64 / number_of_rows;
         reduction += probability_i * probability_i;
     });
     impurity - reduction
@@ -54,7 +54,7 @@ mod tests {
         let number_classes = 1;
         let data = vec![vec![1., 2.], vec![1., 2.], vec![1., 2.]];
         let class_counts = get_class_counts(&vec![0., 0., 0.], number_classes);
-        let gini_result = calculate_gini(&class_counts, data.len() as f32);
+        let gini_result = calculate_gini(&class_counts, data.len() as f64);
         assert_eq!(gini_result, 0.0);
     }
 
@@ -63,7 +63,7 @@ mod tests {
         let number_classes = 3;
         let data = vec![vec![1., 2., 0.], vec![1., 2., 1.], vec![1., 2., 2.]];
         let class_counts = get_class_counts(&vec![0., 1., 2.], number_classes);
-        let gini_result = calculate_gini(&class_counts, data.len() as f32);
+        let gini_result = calculate_gini(&class_counts, data.len() as f64);
         assert_eq!(gini_result, 1.0 - (1.0 / 3.0));
     }
 }
